+++ packages/codegen/src/index.ts
@@ -3,5 +3,9 @@ export * from "./get-new-types"
 export * from "./generate-descriptors"
 export * from "./generate-multiple-descriptors"
 export * from "./generate-docs-descriptors"
-export { knownTypesRepository } from "./known-types"
+export {
+  knownTypes,
+  knownTypesRepository,
+  type KnownTypes,
+} from "./known-types"
 export { generateInkTypes } from "./ink-types"